move event + http handler code to the events package
