+++ services/activitylog/pkg/service/event/l10n/locale/de/LC_MESSAGES/activitylog.po
