+++ services/activitylog/pkg/service/event/l10n/locale/el/LC_MESSAGES/activitylog.po
