+++ services/activitylog/pkg/service/event/l10n/locale/fi/LC_MESSAGES/activitylog.po
