+++ services/activitylog/pkg/service/event/l10n/locale/fr/LC_MESSAGES/activitylog.po
