+++ services/activitylog/pkg/service/event/l10n/locale/hu/LC_MESSAGES/activitylog.po
