+++ services/activitylog/pkg/service/event/l10n/locale/it/LC_MESSAGES/activitylog.po
