+++ services/activitylog/pkg/service/event/l10n/locale/ko/LC_MESSAGES/activitylog.po
