+++ services/activitylog/pkg/service/event/l10n/locale/ru/LC_MESSAGES/activitylog.po
