+++ services/activitylog/pkg/service/event/l10n/locale/zh/LC_MESSAGES/activitylog.po
